Add printf format checking to Show* loggers; fix 17 exposed bugs + ASan OOB
showmsg.h: the Show{Message,Status,SQL,Info,Notice,Warning,Debug,Error,
FatalError} vararg wrappers never had __attribute__((format(printf))),
so -Wall has never checked any of their format strings - the one
remaining place x64 varargs bugs could hide. Adding the attribute
exposed 17 real mismatches, all fixed:

- script.c: 8 ShowWarning calls printed an event-name char* under %d
  (config truncation warnings showed a truncated pointer instead of
  the label).
- mob.c, skill.c: %d directive with NO argument at all (read a
  nonexistent vararg: mob-id / skill-id warnings printed garbage).
- map.c, chrif.c: extra arguments after the format (dropped).
- status.c x2, mob.c, script.c x2, map.c: intptr_t printed as %d
  (post-x64 widened vals/timer data) -> (int) casts.
- pc.c x2 (ARRAYLENGTH/size_t), clif.c (RFIFOREST/size_t) under %d
  -> (int)/(unsigned) casts; mob.c x2: %lu for an unsigned int -> %u.

db.c db_dup_key: duplicate string keys with strnlen, not a fixed
maxlen memcpy - the old code read up to maxlen bytes past the end of
the source allocation (AddressSanitizer flagged it on every script
label strdb_put during boot, reading 50 bytes at the end of str_buf).

Validation: full cluster (login+char+map, 26 scripts) booted under
-fsanitize=address,undefined to completion - 0 ASan errors end to end;
the only UBSan class left is the historical unaligned R/WFIFO buffer
access idiom, harmless on x86-64. Normal rebuild: 0 errors, boots,
OnInit fires on 15 NPCs.

+++ src/map/clif.c
@@ -11492,7 +11492,7 @@ int clif_parse(int fd)
 
 	// filter out invalid / unsupported packets
 	if (cmd > MAX_PACKET_DB || packet_db[packet_ver][cmd].len == 0) {
-		ShowWarning("clif_parse: Received unsupported packet (packet 0x%04x, %d bytes received), disconnecting session #%d.\n", cmd, RFIFOREST(fd), fd);
+		ShowWarning("clif_parse: Received unsupported packet (packet 0x%04x, %d bytes received), disconnecting session #%d.\n", cmd, (int)RFIFOREST(fd), fd);
 		set_eof(fd);
 		return 0;
 	}
